Moar staf

+++ src/io/github/rotemdev/api/Stage.java
@@ -1,10 +1,10 @@
 package io.github.rotemdev.api;
 
 public class Stage {
-	public static final int BLEEDING = -1;
-	public static final int ALPHA = 0;
-	public static final int BETA = 1;
-	public static final int PRE_RELEASE = 2;
-	public static final int RELEASE = 3;
-	public static final int MAJOR_RELEASE = 4;
+	public static final byte BLEEDING = -1;
+	public static final byte ALPHA = 0;
+	public static final byte BETA = 1;
+	public static final byte PRE_RELEASE = 2;
+	public static final byte RELEASE = 3;
+	public static final byte MAJOR_RELEASE = 4;
 }
